05-MCP/project/mcp-weather/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
mcp>=0.1.0